api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
exceptiongroup==1.1.3
fastapi==0.103.2
firebase-admin==6.2.0
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
msgpack==1.0.7
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
uuid==1.30
uvicorn==0.23.2
